# [MPS] Gather sliced inputs to batch norm