@hvaara 2024-08-15
cc @bauerwer for visibility