@bauerwer 2024-08-16
@hvaara Quick check passed on my system. I integrated the Normalization.mm changes into my torch main, built locally and was able to run flux without noise in the image. I ran a few other checkpoints and workflows as well without problems.
![Screenshot 2024-08-16 at 02 03 00](https://github.com/user-attachments/assets/129cc838-6a06-4999-955d-ef5d4dbefdf2)
![Screenshot 2024-08-16 at 02 17 01](https://github.com/user-attachments/assets/c637ac0e-4578-4bc0-8f2c-4ec4916b226b)