@atalman 2024-08-21
@pytorchbot cherry-pick --onto release/2.4 -c critical --fixes https://github.com/pytorch/pytorch/issues/133520